# modules/web-ui/openapi.yaml
openapi: 3.0.3
info:
  title: M6 Web UI Service API (BFF)
  description: |
    **POKER-BRAIN Web UI Service**

    사용자 인터페이스 및 Backend for Frontend (BFF) API입니다.

    **담당**: Frontend Engineer (Frank)
    **모듈 ID**: M6
    **배포**: Cloud Run (us-central1) 또는 Vercel
    **Tech Stack**: Next.js 14 (App Router) + shadcn/ui

    ## 주요 기능
    - 포커 핸드 검색 UI
    - 프록시 영상 미리보기
    - 클립 다운로드 요청 및 상태 조회
    - 즐겨찾기 관리
    - 관리자 대시보드 (Phase 0 진행률)

    ## BFF 패턴
    ```
    Browser (React) → Next.js API Routes (BFF)
        ↓
    BFF → M4 (RAG Search)
    BFF → M5 (Clipping)
    BFF → M3 (Timecode Validation, Admin)
        ↓
    Browser ← JSON Response
    ```

    ## 페이지 구조
    ```
    /                    → 검색 페이지
    /search?q=...        → 검색 결과
    /hand/[hand_id]      → 핸드 상세 + 미리보기
    /downloads           → 내 다운로드 목록
    /favorites           → 즐겨찾기
    /admin               → 관리자 대시보드
    ```

    ## 의존성
    - **Upstream**: M4 (Search), M5 (Clipping), M3 (Admin)
    - **Auth**: Google IAP (자동)

  version: 1.0.0
  contact:
    name: GG Production Frontend Team
    email: aiden.kim@ggproduction.net
servers:
  - url: https://poker-brain-{env}.run.app
    description: Cloud Run deployment
    variables:
      env:
        default: prod
        enum:
          - prod
          - staging
          - dev
  - url: http://localhost:3000
    description: Local development (Next.js)
...
paths:
  /api/search:
    post:
      summary: 포커 핸드 검색 (BFF)
      description: |
        M4 RAG Search Service를 호출하여 검색 결과를 반환합니다.

        **BFF 역할**:
        - M4 API 호출
        - 에러 핸들링
        - 사용자 컨텍스트 추가 (email, favorites)
        - 응답 변환
      operationId: search
      tags:
        - Search
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            example:
              query: "Tom Dwan 블러프"
              filters:
                year_range: [2008, 2024]
              limit: 20
      responses:
        '200':
          description: 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                query_id: "search-20241117-001"
                total_results: 156
                results:
                  - hand_id: "wsop2008_me_d3_h154"
                    summary: "Tom Dwan, J4o, river all-in bluff vs Phil Hellmuth"
                    proxy_url: "https://storage.googleapis.com/..."
                    is_favorite: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 2
        filters:
          type: object
          properties:
            players:
              type: array
              items:
                type: string
            event_name_contains:
              type: string
            year_range:
              type: array
              items:
                type: integer
            pot_size_min:
              type: number
        limit:
          type: integer
          default: 20
    SearchResponse:
      type: object
      properties:
        query_id:
          type: string
        total_results:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/HandSummary'
    HandSummary:
      type: object
      properties:
        hand_id:
          type: string
        summary:
          type: string
        event_name:
          type: string
        timestamp_start:
          type: string
          format: date-time
        proxy_url:
          type: string
          nullable: true
        players:
          type: array
          items:
            type: string
        pot_size_usd:
          type: number
        is_favorite:
          type: boolean
          description: 현재 사용자의 즐겨찾기 여부
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: 내부 서버 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
      description: Next-Auth session cookie (IAP 통합)
security:
  - CookieAuth: []